requirements.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cheroot==10.0.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
gcloud==0.18.3
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jaraco.functools==4.1.0
Jinja2==3.1.5
jwcrypto==1.5.6
MarkupSafe==3.0.2
more-itertools==10.6.0
oauth2client==4.1.3
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyparsing==3.2.1
Pyrebase4==4.8.0
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==75.8.0
six==1.17.0
typing_extensions==4.12.2
urllib3==1.26.20
web.py==0.62
Werkzeug==3.1.3
wheel==0.45.1
